bg(move): Move  __mockData__ to mock folder - Moved mock file to central mock folder [#161313893 Finishes]
